requirements.txt
Django==1.7.1
PyPDF2==1.23
djangorestframework==2.4.3
wsgiref==0.1.2
